Initialize user data when a user is first created
Resolves #49

When a user signs in for the first time, the user document under `Users/<uid>` is created and initialized with the `uid`, `username`, and `profile_picture` fields (although the picture URL may be undefined if the auth method does not provide it).

When a user is deleted, the associated user document is deleted as well.

Automated tests were added to cover these features.