Update the Game document schema
`turn_order` is now a simple array, instead of a map with the turn index as the key.

It seems something changed recently, which required explicitly pulling in firebase-admin 7.0.0 and updating typescript to 3.3.1.

Also updated the `fb-clean` npm script to delegate to `sh/test-cleanup.sh`, which will delete all the Games in addition to the Decks for the two test players.

Resolves #51.