Cloud Functions now set additional initial data.
Updated the hostGameFunction to set `host.uid`, `host.username`, `guests` (which is a map of uid keys to uid and username values), and `status` (which may be one of Pending, In Progress, and Ended).

Updated joinGameFunction to set `guests.<uid>`.

Added lastModified{Game,Player,Card}Function, which updates the base Game document with a `last_modified` field set to the timestamp of the latest modification. This is currently disabled for the Game document as I don't believe we need it (and have not implemented the logic to prevent infinite loops).

Tests updated for the above (save the lastModification functions). I also doubled the timeout for functions as cold starts can be very slow.

Updated firebase-functions to 2.3.1.

Updated engine to node8.

Closes #73.
Closes #78.